Using entity start date instead of trust start during validation when removing beneficiary.
