+++ app/controllers/classofbeneficiary/remove/WhenRemovedController.scala
@@ -20,7 +20,6 @@ import controllers.actions.StandardActionSets
 import forms.DateRemovedFromTrustFormProvider
 import javax.inject.Inject
 import models.{BeneficiaryType, RemoveBeneficiary}
-import navigation.Navigator
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.PlaybackRepository
@@ -28,7 +27,7 @@ import services.TrustService
 import uk.gov.hmrc.play.bootstrap.controller.FrontendBaseController
 import views.html.classofbeneficiary.remove.WhenRemovedView
 
-import scala.concurrent.ExecutionContext
+import scala.concurrent.{ExecutionContext, Future}
 
 class WhenRemovedController @Inject()(
                                        override val messagesApi: MessagesApi,
@@ -44,32 +43,28 @@ class WhenRemovedController @Inject()(
   def onPageLoad(index: Int): Action[AnyContent] = standardActionSets.verifiedForUtr.async {
     implicit request =>
 
-      val form = formProvider.withPrefixAndTrustStartDate("classOfBeneficiary.whenRemoved", request.userAnswers.whenTrustSetup)
-
       trust.getUnidentifiedBeneficiary(request.userAnswers.utr, index).map {
         beneficiary =>
+          val form = formProvider.withPrefixAndEntityStartDate("classOfBeneficiary.whenRemoved", beneficiary.entityStart)
           Ok(view(form, index, beneficiary.description))
       }
-
-
   }
 
   def onSubmit(index: Int): Action[AnyContent] = standardActionSets.verifiedForUtr.async {
     implicit request =>
 
-      val form = formProvider.withPrefixAndTrustStartDate("classOfBeneficiary.whenRemoved", request.userAnswers.whenTrustSetup)
-
-      form.bindFromRequest().fold(
-        formWithErrors =>{
-          trust.getUnidentifiedBeneficiary(request.userAnswers.utr, index).map {
-            beneficiary =>
-              BadRequest(view(formWithErrors, index, beneficiary.description))
-          }
-        },
-        value =>
-          trustService.removeBeneficiary(request.userAnswers.utr, RemoveBeneficiary(BeneficiaryType.ClassOfBeneficiary, index, value)).map(_ =>
-            Redirect(controllers.routes.AddABeneficiaryController.onPageLoad())
+      trust.getUnidentifiedBeneficiary(request.userAnswers.utr, index).flatMap {
+        beneficiary =>
+          val form = formProvider.withPrefixAndEntityStartDate("classOfBeneficiary.whenRemoved", beneficiary.entityStart)
+          form.bindFromRequest().fold(
+            formWithErrors => {
+              Future.successful(BadRequest(view(formWithErrors, index, beneficiary.description)))
+            },
+            value =>
+              trustService.removeBeneficiary(request.userAnswers.utr, RemoveBeneficiary(BeneficiaryType.ClassOfBeneficiary, index, value)).map(_ =>
+                Redirect(controllers.routes.AddABeneficiaryController.onPageLoad())
+              )
           )
-      )
+      }
   }
 }
