+++ test/views/classofbeneficiary/remove/WhenRemovedViewSpec.scala
@@ -31,7 +31,7 @@ class WhenRemovedViewSpec extends QuestionViewBehaviours[LocalDate] {
   val index = 0
   val description= "Some Description"
 
-  override val form: Form[LocalDate] = new DateRemovedFromTrustFormProvider().withPrefixAndTrustStartDate(messageKeyPrefix, LocalDate.now())
+  override val form: Form[LocalDate] = new DateRemovedFromTrustFormProvider().withPrefixAndEntityStartDate(messageKeyPrefix, LocalDate.now())
 
   "whenRemoved view" must {
