+++ test/views/individualbeneficiary/remove/WhenRemovedViewSpec.scala
@@ -30,7 +30,7 @@ class WhenRemovedViewSpec extends QuestionViewBehaviours[LocalDate] {
   val index = 0
   val name = "Name"
 
-  override val form: Form[LocalDate] = new DateRemovedFromTrustFormProvider().withPrefixAndTrustStartDate(messageKeyPrefix, LocalDate.now())
+  override val form: Form[LocalDate] = new DateRemovedFromTrustFormProvider().withPrefixAndEntityStartDate(messageKeyPrefix, LocalDate.now())
 
   "whenRemoved view" must {
